Changed the split position precision to double as issues were occurring where the AABB could be inside out (i.e. BB is smaller than AA) caused by conversion from doubles (in the coordinates) to floats.
Also added a few checks to help in case the issues come up again. Will want to remove the checks when the code is stable to increase performance.

+++ fastKdTree/kdTreeNode.hpp
@@ -161,7 +161,7 @@ namespace kdTree {
         
         AABB BVforAllTris();
         void medianSplit(kdTreeNode **left, kdTreeNode **rght);
-        void split(int dim, float pos, kdTreeNode *left, kdTreeNode *rght) ;
+        void split(int dim, double pos, kdTreeNode *left, kdTreeNode *rght) ;
         
     };
